# docs/Entrega 4/api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Beauty Marketplace API",
    "description": "Documentacao OpenAPI da Entrega 4: catalogo, carrinho, checkout, pedidos e IA.",
    "version": "v1"
  },
  "paths": {
...
    "/api/produtos": {
      "get": {
        "tags": [
          "ApiProdutos"
        ],
        "summary": "Lista produtos do catalogo com filtros de beleza.",
        "description": "Exemplo: GET /api/produtos?pele=Oleosa&vegano=true&precoMax=120",
        "parameters": [
          {
            "name": "busca",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoria",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pele",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cabelo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "marca",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vegano",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "precoMax",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProdutoResumoResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/produtos/{id}": {
      "get": {
        "tags": [
          "ApiProdutos"
        ],
        "summary": "Consulta os detalhes de um produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoDetalheResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/produtos/{id}/avaliacoes": {
      "get": {
        "tags": [
          "ApiProdutos"
        ],
        "summary": "Lista avaliacoes de um produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvaliacaoResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/produtos/{id}/recomendacoes": {
      "get": {
        "tags": [
          "ApiProdutos"
        ],
        "summary": "Recomenda produtos relacionados ao item informado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProdutoRecomendadoResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvaliacaoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "produtoId": {
            "type": "integer",
            "format": "int32"
          },
          "usuarioEmail": {
            "type": "string",
            "nullable": true
          },
          "nota": {
            "type": "integer",
            "format": "int32"
          },
          "comentario": {
            "type": "string",
            "nullable": true
          },
          "midiaUrl": {
            "type": "string",
            "nullable": true
          },
          "tipoPele": {
            "type": "string",
            "nullable": true
          },
          "tipoCabelo": {
            "type": "string",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProdutoDetalheResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "descricao": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "preco": {
            "type": "number",
            "format": "double"
          },
          "estoque": {
            "type": "integer",
            "format": "int32"
          },
          "imagemUrl": {
            "type": "string",
            "nullable": true
          },
          "tipoPele": {
            "type": "string",
            "nullable": true
          },
          "tipoCabelo": {
            "type": "string",
            "nullable": true
          },
          "curvaturaCachos": {
            "type": "string",
            "nullable": true
          },
          "tom": {
            "type": "string",
            "nullable": true
          },
          "acabamento": {
            "type": "string",
            "nullable": true
          },
          "vegano": {
            "type": "boolean"
          },
          "composicao": {
            "type": "string",
            "nullable": true
          },
          "lojistaId": {
            "type": "integer",
            "format": "int32"
          },
          "lojista": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProdutoRecomendadoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "preco": {
            "type": "number",
            "format": "double"
          },
          "justificativa": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProdutoResumoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "preco": {
            "type": "number",
            "format": "double"
          },
          "estoque": {
            "type": "integer",
            "format": "int32"
          },
          "imagemUrl": {
            "type": "string",
            "nullable": true
          },
          "tipoPele": {
            "type": "string",
            "nullable": true
          },
          "tipoCabelo": {
            "type": "string",
            "nullable": true
          },
          "tom": {
            "type": "string",
            "nullable": true
          },
          "vegano": {
            "type": "boolean"
          },
          "lojista": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}